app/backend/requirements.txt
click==8.1.1
Flask==2.1.1
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
protobuf==3.19.4
six==1.16.0
Werkzeug==2.1.0
zipp==3.7.0
